refactor(header and dropdownmenu): create component dropdownMenu by taking it out from header
