+++ src/components/Header/header.scss
@@ -12,49 +12,50 @@
   align-items: center;
   justify-content: center;
   padding: 2px 5px;
+}
 
-  &__menu {
-    width: 54px;
-    height: 47px;
-    margin: 12px, 0px, 0px, 8px;
-    cursor: pointer;
+.header__menu {
+  width: 54px;
+  height: 47px;
+  margin: 12px, 0px, 0px, 8px;
+  cursor: pointer;
 
-    &:hover {
-      width: 60px;
-      height: 55px;
-    }
+  &:hover {
+    width: 60px;
+    height: 55px;
   }
+}
 
-  &__right-block {
-    display: flex;
-    height: 30px;
-    align-items: center;
-    justify-content: center;
-    background: $transparentGreykBackground;
-    margin-right: 0px;
-    margin-left: auto;
-    border-radius: $buttonBorderRadius;
-  }
+.header__right-block {
+  display: flex;
+  height: 30px;
+  align-items: center;
+  justify-content: center;
+  background: $transparentGreykBackground;
+  margin-right: 0px;
+  margin-left: auto;
+  border-radius: $buttonBorderRadius;
 }
 
 .language-toggle {
   display: flex;
   margin-right: 2px;
   align-items: center;
+}
 
-  &__icon {
-    color: $white;
-    font-size: $fontsize-20;
-    font-weight: $bold;
-  }
+.language-toggle__icon {
+  color: $white;
+  font-size: $fontsize-20;
+  font-weight: $bold;
 }
 
 .header-icons {
   display: flex;
   justify-content: right;
   align-items: center;
+}
 
-  &__item {
+  .header-icons__item {
     width: auto;
     height: 20px;
     margin-right: 2px;
@@ -64,56 +65,10 @@
       height: 30px;
     }
   }
-}
 
 .btn-light-theme {
   font-size: $fontsize-14;
   padding: 4px 2px;
-}
-
-.user-name {
-  margin: 0 3px 0 5px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  color: $white;
-}
-
-.dropdown-toggle {
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  cursor: pointer;
-  color: $white;
-
-  @media (max-width: $mobile-breakpoint) {
-    margin-bottom: 2px;
-  }
-}
-
-.dropdown-menu {
-  background-color: $white;
-  padding: 0;
-  margin-top: 10px;
-  border-radius: $buttonBorderRadius;
-}
-.dropdown-item {
-  background-color: $white;
-  color: $lightThemeAccept;
-  margin: 2px;
-  font-size: $fontsize-10;
-
-  &:hover {
-    padding-left: 50px;
-    border-bottom: 3px solid $lightThemeAccept;
-  }
-}
-.removedecorations,
-.removedecorations:hover {
-  all: unset;
 }
 
 @media (max-width: $mobile-breakpoint) {
