+++ src/components/MainLayout/MainLayout.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Header from "../../containers/headerContainer";
+import Header from "../Header/header";
 import Footer from "../footer/footer";
 import AlertContainer from "../../containers/alert";
 import "./main-layout.css";
